- Separate todo and todo input into separate folders with .js and .css file
