+++ src/components/todo/TodoComponent.js
@@ -1,6 +1,7 @@
-import React, { Component } from 'react';
+import './todo-styles.css';
+import * as React from 'react';
 
-export class Todo extends Component {
+export default class Todo extends React.Component {
   render () {
     const { text, onClickDelete } = this.props;
     return (
@@ -8,6 +9,6 @@ export class Todo extends Component {
         {text}
         <span onClick={onClickDelete}>&times;</span>
       </div>
-    );
+    )
   }
 }
